rename repositories/openai to repositories/llm and update API route to /llm
